requirements.txt
faker==40.5.1
pandas==3.0.1
numpy==2.4.2
jupyter==1.1.1
jupyterlab==4.5.5
matplotlib==3.10.8
seaborn==0.13.2
fastapi==0.133.0
uvicorn==0.41.0
python-multipart==0.0.22
jinja2==3.1.6
python-dotenv==1.2.1
sqlalchemy==2.0.47
psycopg2-binary==2.9.11
